requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
django-cors-headers==3.13.0